requirements.txt
Flask==2.3.3
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
flask-jwt-extended==4.4.4
passlib[bcrypt]==1.7.5
gunicorn==20.1.0
gtts
flask_cors